Enhanced_Telemetry_Agent/requirements.txt
Flask
flask-httpauth
pymongo[srv]
PyQt5
requests
grpcio
protobuf
pyyaml
influxdb_client
